requirements.txt
streamlit
tiktoken
PyMuPDF==1.22.3
pdfplumber
openai
qdrant-client
azure-storage-blob